cdk/requirements.txt
aws-cdk.aws-apigateway==1.98.0
aws-cdk.aws-ecr==1.98.0
aws-cdk.aws-lambda==1.98.0
aws-cdk.aws-s3==1.98.0
